[ProtocolLib]: minor directory refactor.
